\name{mapDHf}
\alias{mapDHf}
\docType{data}
\title{An unconstructed marker set for a doubled haploid wheat population}
\description{An unconstructed marker set for a doubled
  haploid wheat population in the form of a constructed \pkg{qtl} object.
}
\usage{data(mapDHf)}
\format{This data is the unconstructed version of \code{\link{mapDH}}
  and consists of 599 markers genotyped on 218 individuals. 584 markers
  are from the orignal map with an additonal 12 co-located markers and
  3 slightly distorted markers. The data is in a \code{data.frame} format
  with genotypes in columns and randomly assorted markers in rows. See
  \code{\link{mstmap.data.frame}} documentation for more details on the format
  of this object.
}
\examples{

data(mapDHf, package = "ASMap")

}
\keyword{datasets}