\name{mapDHu}
\alias{mapDHu}
\docType{data}
\title{An unconstructed linkage map for a doubled haploid wheat population}
\description{An unconstructed linkage map for a doubled
  haploid wheat population in the form of a constructed R/qtl object.
}
\usage{data(mapDHu)}
\format{This data is the unconstructed version of \code{\link{mapDH}}
  and consists of 620 markers spanning one large linkage group genotyped
  on 218 individuals. 584 markers are from the original map and have
  been randomized within the linkage group. In addition, the
  unconstructed map contains 12 markers that co-locate with other
  markers, 12 markers with excessive segregation distortion and 12
  markers with excessive missing values. These extra markers have been
  artificially placed in the object for the purpose of highlighting
  features of the package.  
}
\examples{

data(mapDHu, package = "ASMap")

}
\keyword{datasets}